+++ AcopioAPIs/DTOs/Recojo/RecojoInsertDto.cs
@@ -6,7 +6,7 @@ namespace AcopioAPIs.DTOs.Recojo
     {
         public DateOnly RecojoFechaInicio { get; set; }
         public DateOnly RecojoFechaFin { get; set; }
-        public int RecojoCamionesCantidad { get; set; }
+        public decimal RecojoCamionesCantidad { get; set; }
         public decimal RecojoCamionesPrecio { get; set; }
         public int RecojoDiasCantidad { get; set; }
         public decimal RecojoDiasPrecio { get; set; }
